fixed issue #12, added orientation updates, draggable nodes w/ r-click
Patched the bug mention in issue #12.
Orientation is now displayed on a re-arranged side canvas with the ability to update it from values placed into the inputs.
Right click and hold to drag a node around on the canvas, position is automatically updated in the tmapdata and node connections are redrawn once dropped.