create new file from just pgm, synced git api, updated help txt
Added option to create a new project from just a PGM file without needing any YAML or TMAP data. Added a connection request through the git api for the repos latest release, should display version number and latest release notes. Added a save button in addition to the save as (one prompts the user for location the other automatically overwrites the existing file). Fixed a bug with the recent files where it would only ever open the last file in the list from the buttons. Added a new class for the popup window that gets the users YAML data when starting a fresh project. Updated the help text file with some missing features and the requirements.